requirements.txt
azure-functions==1.20.0
msal==1.32.0
requests==2.32.3
ruff==0.12.4
aiohttp==3.12.14
aiosqlite==0.21.0
python-dotenv==1.0.0
backoff==2.2.1
pytz==2024.1
